refactor(repo): create `Repo::get_head_oid`
