+++ src/commands/smartlog.rs
@@ -21,7 +21,7 @@ use crate::core::metadata::{
     RelativeTimeProvider,
 };
 use crate::core::repo::Repo;
-use crate::util::{get_branch_oid_to_names, get_db_conn, get_head_oid, get_main_branch_oid};
+use crate::util::{get_branch_oid_to_names, get_db_conn, get_main_branch_oid};
 
 /// Split fully-independent subgraphs into multiple graphs.
 ///
@@ -273,7 +273,7 @@ pub fn smartlog() -> anyhow::Result<()> {
     let merge_base_db = MergeBaseDb::new(&conn)?;
     let event_log_db = EventLogDb::new(&conn)?;
     let event_replayer = EventReplayer::from_event_log_db(&event_log_db)?;
-    let head_oid = get_head_oid(&repo)?;
+    let head_oid = repo.get_head_oid()?;
     let main_branch_oid = get_main_branch_oid(&repo)?;
     let branch_oid_to_names = get_branch_oid_to_names(&repo)?;
     let graph = make_graph(
