+++ src/core/repo.rs
@@ -9,6 +9,7 @@
 //! codebase.
 //! - To collect some different helper Git functions.
 
+use std::borrow::{Borrow, BorrowMut};
 use std::path::Path;
 
 use anyhow::Context;
@@ -35,6 +36,18 @@ impl std::ops::DerefMut for Repo {
     }
 }
 
+impl Borrow<git2::Repository> for Repo {
+    fn borrow(&self) -> &git2::Repository {
+        &self.repo
+    }
+}
+
+impl BorrowMut<git2::Repository> for Repo {
+    fn borrow_mut(&mut self) -> &mut git2::Repository {
+        &mut self.repo
+    }
+}
+
 impl From<git2::Repository> for Repo {
     fn from(repo: git2::Repository) -> Self {
         Repo { repo }
@@ -42,8 +55,9 @@ impl From<git2::Repository> for Repo {
 }
 
 impl Repo {
+    /// Get the Git repository associated with the given directory.
     #[context("Getting Git repository for directory: {:?}", &path)]
-    fn from_dir(path: &Path) -> anyhow::Result<Self> {
+    pub fn from_dir(path: &Path) -> anyhow::Result<Self> {
         let repository = git2::Repository::discover(path).map_err(wrap_git_error)?;
         Ok(repository.into())
     }
@@ -53,5 +67,22 @@ impl Repo {
     pub fn from_current_dir() -> anyhow::Result<Self> {
         let path = std::env::current_dir().with_context(|| "Getting working directory")?;
         Repo::from_dir(&path)
+    }
+
+    /// Get the OID for the repository's `HEAD` reference.
+    #[context("Getting HEAD OID for repository at : {:?}", self.repo.path())]
+    pub fn get_head_oid(&self) -> anyhow::Result<Option<git2::Oid>> {
+        let head_ref = match self.repo.head() {
+            Ok(head_ref) => Ok(head_ref),
+            Err(err)
+                if err.code() == git2::ErrorCode::NotFound
+                    || err.code() == git2::ErrorCode::UnbornBranch =>
+            {
+                return Ok(None)
+            }
+            Err(err) => Err(err),
+        }?;
+        let head_commit = head_ref.peel_to_commit()?;
+        Ok(Some(head_commit.id()))
     }
 }
